pyproject.toml
[project]
name = "rag-llm"
version = "0.1.0"
description = ""
authors = [
    {name = "Haider Qaizar Hussain",email = "haiderqaizar3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "chromadb[pysqlite3] (>=0.6.3)",
    "fastapi (>=0.115.7,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "langchain-text-splitters (>=0.3.5,<0.4.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "langchain-chroma (>=0.2.1,<0.3.0)",
    "langchain-openai (>=0.3.2,<0.4.0)",
    "gradio (>=5.13.2,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "uuid (>=1.30,<2.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pypdf (>=5.2.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
